requirements.txt
django==3.0.8
djangorestframework==3.11.0
django-environ==0.4.5
pytest==5.4.3
pytest-django==3.9.0
pytest-mock==3.1.1
Faker==4.1.1
mixer==6.1.3